Fixed missing level_83: Now level_84 becomes level_83.
As about other thing stated in #123733:
Yes, levels 47 and 81 have a similar molecule name, but
that's not a bug, because it IS a same molecule, but it layed
out differently in each of these levels which produces completely
different gaming experience for each of them ;).

Slider 'jumping' is fixed too.

BUG: 123733

svn path=/trunk/KDE/kdegames/katomic/; revision=602822