Well, here's the "KAtomic highscores rethinked" :-)
Instead of showing KScoreDialog after every level
(which isa little annoying imho), I wrote a simple 
QGraphicsItem that behaves like a passive popup
which shows a message to the user for specified amount
of time and then hides itself.

Its has animated showing/hiding, nice icon and text.

I'd be grateful for any comments/suggestions.
(To see it in action you have to complete some level
and make a new highscore - shouldn't be hard to do ;)

If any other games find this useful we may even move
it to libkdegames. And then KGameCanvas variant should
be written too.

svn path=/trunk/KDE/kdegames/katomic/; revision=675468